restructure the menu and actionbar ref #55258
